feat(Dashboard): added delete movie

+++ backend/src/modules/movie.js
@@ -1,44 +1,33 @@
 const express = require('express')
-const Movie = require('../models/moviesSchema')
-const missingFields = require('../helpers/missingFields')
 const multer = require('multer')
+const Movie = require('../models/moviesSchema')
+const ResponseWithoutPayload = require('../classes/responseWithoutPayload')
+const ResponseDashboardData = require('../classes/responseDashboardData')
+const ResponseUserData = require('../classes/responseUserData')
 
 const router = express.Router()
 
 const upload = multer({ dest: 'public/posters' })
 async function createMovie(req, res) {
-  upload.single('poster')(req, res, async function (err) {
-    if (err) {
-      return res.status(500).send({ status: 500, message: 'File upload failed' })
+  upload.single('poster')(req, res, async function () {
+    if (req.clientResponseError) {
+      return res.status(req.clientResponseError.status).send(req.clientResponseError)
     }
+
     const posterPath = req.file.path
 
     const { title, year, releaseDate, boxOffice, country, production, actors, directors, writers, rating, imdbRating, genre, description, duration, isRecommended } = req.body
-    const errMessages = []
-    missingFields(title, 'Title', errMessages)
-    missingFields(year, 'Year', errMessages)
-    missingFields(releaseDate, 'Release Date', errMessages)
-    missingFields(boxOffice, 'Box Office', errMessages)
-    missingFields(country, 'Country', errMessages)
-    missingFields(production, 'Production', errMessages)
-    missingFields(actors, 'Actors', errMessages)
-    missingFields(directors, 'Directors', errMessages)
-    missingFields(writers, 'Writers', errMessages)
-    missingFields(rating, 'Rating', errMessages)
-    missingFields(imdbRating, 'IMDb Rating', errMessages)
-    missingFields(genre, 'Genre', errMessages)
-    missingFields(description, 'Description', errMessages)
-    missingFields(duration, 'Duration', errMessages)
-    missingFields(isRecommended, 'Recommended', errMessages)
-
-    if (errMessages.length > 0) {
-      return res.status(400).send({
-        status: 400, message:
-          `Missing fields: ${errMessages.join(', ')}`
-      })
-    }
 
     try {
+      const { accessToken, tokenValid } = req.userData.token
+
+      if (!accessToken) {
+        return res.status(401).send(new ResponseWithoutPayload(401, 'Access token is required'))
+      }
+
+      if (!tokenValid) {
+        return res.status(401).send(new ResponseWithoutPayload(401, 'Invalid access token'))
+      }
       const movie = new Movie({
         title,
         year,
@@ -57,10 +46,11 @@ async function createMovie(req, res) {
         description,
         isRecommended
       })
+
       await movie.save()
-      res.status(201).send({ status: 201, message: 'Movie created successfully', movie })
+      res.status(201).send(new ResponseDashboardData(201, 'Movie created successfully', movie))
     } catch (error) {
-      res.status(500).send({ status: 500, message: 'Internal Server Error' })
+      res.status(500).send(new ResponseWithoutPayload(500, 'Internal Server Error'))
     }
   })
 }
@@ -69,37 +59,46 @@ async function getMovieById(req, res) {
   const id = req.params.id
 
   if (!id) {
-    return res.status(400).send({ status: 400, message: 'Movie ID is required' })
+    return res.status(400).send(new ResponseWithoutPayload(400, 'Movie ID is required'))
   }
 
   try {
     const movie = await Movie.findById(id)
     if (movie) {
-      res.status(200).send({ status: 200, message: 'Success', movie })
+      res.status(200).send(new ResponseDashboardData(200, 'Success', movie))
     } else {
-      res.status(404).send({ status: 404, message: 'Movie not found' })
+      res.status(404).send(new ResponseWithoutPayload(404, 'Movie not found'))
     }
   } catch (error) {
-    res.status(500).send({ status: 500, message: error.message })
+    res.status(500).send(new ResponseWithoutPayload(500, 'Internal Server Error'))
   }
 }
 
 async function deleteMovieById(req, res) {
-  const id = req.params.id
+  const { id } = req.params
+  const { accessToken, tokenValid } = req.userData.token
+
+  if (!accessToken) {
+    return res.status(401).send(new ResponseUserData(401, 'Access token is required', null))
+  }
+
+  if (!tokenValid) {
+    return res.status(401).send(new ResponseUserData(401, 'Invalid access token', null))
+  }
 
   if (!id) {
-    return res.status(400).send({ status: 400, message: 'Movie ID is required' })
+    return res.status(400).send(new ResponseWithoutPayload(400, 'Movie ID is required'))
   }
 
   try {
     const movie = await Movie.findByIdAndDelete(id)
     if (movie) {
-      res.status(200).send({ status: 200, message: 'Movie deleted successfully' })
+      res.status(200).send(new ResponseWithoutPayload(200, 'Movie deleted successfully'))
     } else {
-      res.status(404).send({ status: 404, message: 'Movie not found' })
+      res.status(404).send(new ResponseWithoutPayload(404, 'Movie not found'))
     }
   } catch (error) {
-    res.status(500).send({ status: 500, message: error.message })
+    res.status(500).send(new ResponseWithoutPayload(500, 'Internal Server Error'))
   }
 }
